Revised Da's FileOp methods to be clearer, simpler, and much more heavily commented.
Fixed up some English.

Created working tests (largely to make sure my changes didn't break anything).

Fixed a terribly misconception I had in the creation of FileOp.convertPath(). It was
returning null if the file didn't exist. Obviously that's no good for the initial copy.
Removing that makes the function much more useful. Also used this function in some
of Da's implementation, which cut the code complexity a bit.

Standardized the useage of Control to access the logger and live/backup directory. It
was a bit all over the place and incompatible with the rest of the code. Now it's compatible.

Fixed a test that didn't properly clean up after itself (which caused a different test to
fail, ugh).

TesterVisual ignores all tests by default, so it's simply a matter of turning on the tests
you want to use. Handy since those tests can only be verified manually.