Added support for changing the current directory via the location bar.
Needed an unusual hook to catch the case where the entered path ends with
".." (but not "/.." or "\.."), because apparently that's a valid directory,
yet later down the road, when calling listFiles() on that path, it's no
longer valid. Even weirder, adding ".." to the end of a directory in Windows
will go into that directory, but Linux treats it as a whole different directory
(as would be expected). Seriously, I have no idea what's with that.