Somehow, we forgot to store the size of revisions in the revisions database.
I have now added the necessary changes, which include the DbManager creating
the table with a column for revision size, getting and setting revisions takes
in a size, and the FileHistory now uses the size that it takes in (we remembered
the need to pass the file size in there, but nowhere else).

As well, the GUI displays the revision in its table, and the table includes
a column for the file size.

This shouldn't require changes to the handlers, as they already passed the revision
size in (although it wasn't doing anything).

This change is NOT BACKWARDS COMPATIBLE. The databases created with earlier versions
of the program will have to be re-created (delete the backup_location file in
trunk/main and go through the first start wizard again).