Got FileOp.createPatch() (previously createDiff()) and FileOp.applyPatch() (previously applyDiff()) working.
I believe that Java-diff-utils was bugged. The parse diff
method was returning an empty list. Removed the library
in question and switched to Google's Diff Match and Patch,
which worked like a charm, and best of all, offers the
functionality to pretty-print our diff. Instead of a
traditional JAR file, this library is simply bundled with
the direct source code, which allows easy access to
documentation, which is losted in compilation.

The FileOp.fileToList() method is now unnecessary and
was removed.

Created a test for our apply and create diff methods.

Removed a bugged test from the old Watcher behavior.